fix(plugins): enforce manifest setting bounds server-side
descriptor_for hard-coded ValidationKind::None, so a plugin Number widget's
min/max and a Select widget's option set were advisory UI metadata only: a web
PATCH could write any integer or any string straight through validate_patch
into config. Derive validation from the widget instead: Number ->
ValidationKind::RangeI64 (new signed-range variant; an unbounded number still
pins to an integer), Select -> ValidationKind::OneOf (new closed-set variant).

SettingWidget::Number bounds are now i64 instead of f64, so a fractional bound
like min = 0.5 is a loud TOML type error at parse rather than a silent
truncation to 0, and widget_for no longer does a lossy  cast.

Addresses review comments from @jerome-benoit on #2090.

+++ aoe-plugin-api/src/manifest.rs
@@ -125,8 +125,16 @@ pub struct SettingContribution {
 pub enum SettingWidget {
     Toggle,
     Text,
-    Number { min: Option<f64>, max: Option<f64> },
-    Select { options: Vec<String> },
+    /// Integer input with optional inclusive bounds. Bounds are `i64`, matching
+    /// the host's integer settings; a fractional bound (`min = 0.5`) is a loud
+    /// TOML type error at parse rather than a silently truncated `0`.
+    Number {
+        min: Option<i64>,
+        max: Option<i64>,
+    },
+    Select {
+        options: Vec<String>,
+    },
 }
 
 /// Override of another setting's default, resolved by priority.
